# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SearchResultType = typeof SearchResultType[keyof typeof SearchResultType];


export const SearchResultType = {
  page: 'page',
  database: 'database',
} as const;

export interface SearchResult {
  id: string;
  title: string;
  type: SearchResultType;
  url?: string;
  createdAt?: string;
}

export interface WorkspaceOverview {
  totalPages: number;
  totalDatabases: number;
  recentItems?: SearchResult[];
}

export type NotionDatabaseProperties = {[key: string]: { [key: string]: unknown }};

export interface NotionDatabase {
  id: string;
  title: string;
  url?: string;
  createdAt?: string;
  properties: NotionDatabaseProperties;
}

export type NotionEntryProperties = { [key: string]: unknown };

export interface NotionEntry {
  id: string;
  /** @nullable */
  title?: string | null;
  url?: string;
  createdAt?: string;
  properties?: NotionEntryProperties;
}

export type EntryUpdateProperties = { [key: string]: unknown };

export interface EntryUpdate {
  properties?: EntryUpdateProperties;
}

export type NotionPageContentItemChildrenItem = { [key: string]: unknown };

export type NotionPageContentItem = {
  type?: string;
  /** @nullable */
  text?: string | null;
  children?: NotionPageContentItemChildrenItem[];
};

export interface NotionPage {
  id: string;
  title: string;
  url?: string;
  createdAt?: string;
  content?: NotionPageContentItem[];
}

export type SearchNotionParams = {
q: string;
};