# lib/api-zod/src/generated/types/notionDatabase.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotionDatabaseProperties } from './notionDatabaseProperties';

export interface NotionDatabase {
  id: string;
  title: string;
  url?: string;
  createdAt?: string;
  properties: NotionDatabaseProperties;
}