# lib/api-zod/src/generated/types/notionEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotionEntryProperties } from './notionEntryProperties';

export interface NotionEntry {
  id: string;
  /** @nullable */
  title?: string | null;
  url?: string;
  createdAt?: string;
  properties?: NotionEntryProperties;
}