# lib/api-zod/src/generated/types/notionPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotionPageContentItem } from './notionPageContentItem';

export interface NotionPage {
  id: string;
  title: string;
  url?: string;
  createdAt?: string;
  content?: NotionPageContentItem[];
}