# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResultType } from './searchResultType';

export interface SearchResult {
  id: string;
  title: string;
  type: SearchResultType;
  url?: string;
  createdAt?: string;
}