# lib/api-zod/src/generated/types/workspaceOverview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResult } from './searchResult';

export interface WorkspaceOverview {
  totalPages: number;
  totalDatabases: number;
  recentItems?: SearchResult[];
}